+++ Scripts/Core/Rendering/BaseCompositorEffect.cs
@@ -51,7 +51,7 @@ public abstract partial class BaseCompositorEffect : CompositorEffect, ISerializ
 
 
 		if (name == PropertyName.RenderingDevice) {
-			property["usage"] = (int)(property["usage"].As<PropertyUsageFlags>() & ~PropertyUsageFlags.Storage);
+			property["usage"] = (long)(property["usage"].As<PropertyUsageFlags>() & ~PropertyUsageFlags.Storage);
 		}
 	}
 }
